Improve trackball/rate/glider/matrix safety and performance: widen wheel buffer, standardize millis types, optimize glider math, use wait_us

+++ clockworkpi/uconsole/timeout.h
@@ -4,7 +4,7 @@
 #define CUTOFF_MS 1000
 
 typedef uint16_t timeout_t;
-typedef uint8_t millis_t;
+typedef uint16_t millis_t;
 
 timeout_t timeout_update(timeout_t, millis_t);
 bool timeout_get(timeout_t);
